EmuStream now has a copy-constructor and operator overload on operator=
This was made to make it possible to use the class with std::vectors and std::map.
std::map now works as intended, and even though there are no visual representation, you can now load all files from a folder, and probably a single file (untested), with validation and logging.