You can now add all tracks to Playlist, but beware of many files (#74)
If you add many thousands of files with many tracks, the lists will be enormous, and this will kill the performance. I Tested this with 31000 files containing a total of 636471 tracks (which would be the exact entry in the list). This would case major performance issues. Treelists should be improved with paging.